+++ spec/components/contextual_breadcrumbs_spec.rb
@@ -144,54 +144,15 @@ describe "ContextualBreadcrumbs", type: :view do
     assert_select "a", text: "EU Withdrawal Act 2018 statutory instruments"
   end
 
-  it "renders parent finder breadcrumb if content has a finder linked and taxon is prioritised" do
-    content_item = example_document_for("guide", "guide-with-facet-groups")
-    render_component(content_item: content_item, prioritise_taxon_breadcrumbs: true)
-
-    assert_select "a", text: "Home"
-    assert_select "a", text: "EU Withdrawal Act 2018 statutory instruments"
-  end
-
-  it "renders inverse parent finder breadcrumb" do
-    content_item = example_document_for("guide", "guide-with-facet-groups")
-    render_component(content_item: content_item, prioritise_taxon_breadcrumbs: true, inverse: true)
-    assert_select ".gem-c-breadcrumbs.gem-c-breadcrumbs--inverse"
-  end
-
   it "renders no breadcrumbs if there aren't any" do
     content_item = example_document_for("guide", "guide")
     content_item = remove_mainstream_browse(content_item)
-    content_item = remove_curated_related_item(content_item)
     content_item["links"]["taxons"] = nil
     assert_no_selector(".gem-c-step-nav-header")
     assert_no_selector(".gem-c-breadcrumbs")
   end
 
-  it "renders taxon breadcrumbs even if there are mainstream browse pages if prioritise_taxon_breadcrumbs is true" do
-    content_item = example_document_for("guide", "guide")
-    content_item = set_parent_titles_to_businesses(content_item)
-    content_item = set_live_taxons(content_item)
-    render_component(content_item: content_item, prioritise_taxon_breadcrumbs: true)
-    assert_select "a", text: "Home"
-    assert_no_selector "a", text: "Business and self-employed"
-    assert_no_selector "a", text: "Licences and licence applications"
-    assert_select "a", text: "School curriculum"
-    assert_select "a", text: "Education, training and skills"
-  end
-
-  it "renders mainstream browse pages if prioritise_taxon_breadcrumbs is false and there are live taxons" do
-    content_item = example_document_for("guide", "guide")
-    content_item = set_parent_titles_to_businesses(content_item)
-    content_item = set_live_taxons(content_item)
-    render_component(content_item: content_item, prioritise_taxon_breadcrumbs: false)
-    assert_select "a", text: "Home"
-    assert_select "a", text: "Business and self-employed"
-    assert_select "a", text: "Licences and licence applications"
-    assert_no_selector "a", text: "School curriculum"
-    assert_no_selector "a", text: "Education, training and skills"
-  end
-
-  it "renders mainstream browse pages if prioritise_taxon_breadcrumbs is not passed and are live taxons" do
+  it "renders mainstream browse pages if there some and there are live taxons" do
     content_item = example_document_for("guide", "guide")
     content_item = set_parent_titles_to_businesses(content_item)
     content_item = set_live_taxons(content_item)