+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 ## Unreleased
 
 * Modify Step nav header component to support custom tracking ([PR #1533](https://github.com/alphagov/govuk_publishing_components/pull/1533))
-
+* Remove prioritise_taxon_breadcrumbs argument from contextual breadcrumbs component ([PR #1534](https://github.com/alphagov/govuk_publishing_components/pull/1534))
 
 ## 21.53.0